refactor: decompose array/mod.rs (3,150 lines) into focused submodules
Split the monolithic mod.rs into 8 focused submodules while preserving
the existing public API. Each submodule contains a separate `impl MxArray`
block with related methods, improving navigability and maintainability.

New submodules:
- creation.rs: array constructors (from_*, zeros, ones, full, linspace, eye, arange)
- ops.rs: arithmetic, math, trig, NaN/Inf checks
- reduction.rs: sum, mean, max, min, argmax, var, std, logsumexp, cumsum, cumprod
- shape.rs: reshape, transpose, slice, concatenate, split, squeeze, pad, sort, stack
- comparison.rs: equal, not_equal, less, greater, logical_and/or/not, where_
- data.rs: eval, shape, dtype, size, to_float32/int32/uint32, item_at_*, copy, astype
- random.rs: random_uniform, random_normal, random_bernoulli, randint, categorical
- memory.rs: free functions for cache/memory management
- tests.rs: all #[cfg(test)] content (~1,565 lines)

mod.rs reduced from 3,150 to 105 lines (struct def, DType, module decls, re-exports).
All 715 Rust tests and 457 TypeScript tests pass unchanged.
